Refactor store_volume() function into a separate module
This will soon be re-used outside of the Docker runner.
