+++ nextstrain/cli/command/shell.py
@@ -7,6 +7,7 @@ check-setup` to check if Docker is installed and works.
 """
 
 from ..runner import docker
+from ..volume import store_volume
 
 
 def register_parser(subparser):
@@ -18,7 +19,7 @@ def register_parser(subparser):
         "directory",
         help    = "Path to pathogen build directory",
         metavar = "<directory>",
-        action  = docker.store_volume("build"))
+        action  = store_volume("build"))
 
     # Runner options
     docker.register_arguments(
